BUG Add missing allfinite() call. Refactor model to models.py
