+++ pymc3/tests/test_starting.py
@@ -3,8 +3,6 @@ import numpy as np
 from pymc3.tuning import starting
 from pymc3 import Model, Uniform, Normal, Beta, Binomial, find_MAP, Point
 from .models import simple_model, non_normal, exponential_beta
-from theano.compile.ops import as_op
-import theano.tensor as tt
 
 def test_accuracy_normal():
     _, model, (mu, _) = simple_model()
@@ -55,16 +53,8 @@ def test_find_MAP_discrete():
 
 
 def test_find_MAP_no_gradient():
-    # as_op without gradient definition
-    @as_op(itypes=[tt.dscalar], otypes=[tt.dscalar])
-    def arbitrary_det(value):
-        return value
-
-    with Model() as model_deterministic:
-        a = Normal('a')
-        b = arbitrary_det(a)
-        c = Normal('obs', mu=b.astype('float64'),
-                   observed=np.array([1, 3, 5]))
+    _, model = simple_arbitrary_det()
+    with model:
         find_MAP()
